+++ src/main/java/de/johni0702/minecraft/bobby/FakeChunk.java
@@ -1,9 +1,11 @@
 package de.johni0702.minecraft.bobby;
 
 import de.johni0702.minecraft.bobby.ext.ChunkLightProviderExt;
+import net.minecraft.block.BlockState;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.WorldRenderer;
 import net.minecraft.nbt.NbtList;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.ChunkSectionPos;
 import net.minecraft.world.Heightmap;
@@ -15,6 +17,7 @@ import net.minecraft.world.chunk.UpgradeData;
 import net.minecraft.world.chunk.WorldChunk;
 import net.minecraft.world.chunk.light.LightingProvider;
 import net.minecraft.world.tick.ChunkTickScheduler;
+import org.jetbrains.annotations.Nullable;
 
 // Fake chunks are of this subclass, primarily so we have an easy way of identifying them.
 public class FakeChunk extends WorldChunk {
@@ -65,5 +68,14 @@ public class FakeChunk extends WorldChunk {
 
     public void setHeightmap(Heightmap.Type type, Heightmap heightmap) {
         this.heightmaps.put(type, heightmap);
+    }
+
+    @Override
+    public @Nullable BlockState setBlockState(BlockPos pos, BlockState state, boolean moved) {
+        // This should never be called for fake chunks, but some server incorrectly send block updates for chunks
+        // they just unloaded, which can then result in a race condition between this update and the background thread
+        // which serializes the chunk.
+        // See https://github.com/Johni0702/bobby/issues/341
+        return null;
     }
 }